+++ pysat/constellations/de2.py
@@ -1,8 +1,13 @@
-import pysat
 """
-Creates a constellation from NASA ICON instrumentation
+Creates a constellation from NASA DE2 instrumentation
+
+.. deprecated:: 2.3.0
+  This routine has been deprecated in pysat 3.0.0, and will be accessible
+  in `pysatNASA.constellations.de2`
+
 """
 
+import pysat
 
 lang = pysat.Instrument(platform='de2', name='lang')
 nacs = pysat.Instrument(platform='de2', name='nacs')