+++ pysat/instruments/de2_wats.py
@@ -52,6 +52,10 @@ J. C. Maurer
 The Dynamics Explorer Wind and Temperature Spectrometer
 Space Sci. Instrum., v. 5, n. 4, p. 417, 1981.
 
+.. deprecated:: 2.3.0
+  This Instrument module has been removed from pysat in the 3.0.0 release and
+  can now be found in pysatNASA (https://github.com/pysat/pysatNASA)
+
 Properties
 ----------
 platform
@@ -73,6 +77,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 
 import functools
+import warnings
 
 import pysat
 from pysat.instruments.methods import nasa_cdaweb as cdw
@@ -105,6 +110,19 @@ download = functools.partial(cdw.download, supported_tags)
 # support listing files currently on CDAWeb
 list_remote_files = functools.partial(cdw.list_remote_files,
                                       supported_tags=supported_tags)
+
+
+def init(self):
+    """Initializes the Instrument object with values
+    """
+
+    warnings.warn(" ".join(["_".join([self.platform, self.name]),
+                            "has been removed from the pysat-managed",
+                            "Instruments in pysat 3.0.0, and now resides in",
+                            "pysatNASA:",
+                            "https://github.com/pysat/pysatNASA"]),
+                  DeprecationWarning, stacklevel=2)
+    return
 
 
 def clean(inst):