+++ pysat/instruments/timed_see.py
@@ -7,6 +7,10 @@ Analysis Web (CDAWeb).
 Supports two options for loading that may be
 specified at instantiation.
 
+.. deprecated:: 2.3.0
+  This Instrument module has been removed from pysat in the 3.0.0 release and
+  can now be found in pysatNASA (https://github.com/pysat/pysatNASA)
+
 Properties
 ----------
 platform
@@ -42,6 +46,7 @@ Warnings
 from __future__ import print_function
 from __future__ import absolute_import
 import functools
+import warnings
 
 import pysat
 from pysat.instruments.methods import nasa_cdaweb as cdw
@@ -77,6 +82,19 @@ list_remote_files = functools.partial(cdw.list_remote_files,
 # support load routine
 # use the default CDAWeb method
 load = functools.partial(cdw.load, fake_daily_files_from_monthly=True)
+
+
+def init(self):
+    """Initializes the Instrument object with values
+    """
+
+    warnings.warn(" ".join(["_".join([self.platform, self.name]),
+                            "has been removed from the pysat-managed",
+                            "Instruments in pysat 3.0.0, and now resides in",
+                            "pysatNASA:",
+                            "https://github.com/pysat/pysatNASA"]),
+                  DeprecationWarning, stacklevel=2)
+    return
 
 
 def clean(inst):