+++ pysat/tests/test_constellation.py
@@ -369,3 +369,28 @@ class TestDeprecation():
         assert str(war[0].message).find(warn_msg) >= 0
 
         return
+
+    def test_deprecation_constellations(self):
+        """Test the deprecation of migrated constellation objects"""
+
+        # Define the deprecated attributes that are always defined
+        self.warn_msg = "This constellation has been removed from"
+        del self.in_kwargs['instruments']
+
+        # Catch the warnings
+        const_names = ['de2', 'icon']
+        for const_name in const_names:
+            with warnings.catch_warnings(record=True) as war:
+                pysat.Constellation(name=const_name)
+
+            # Ensure the minimum number of warnings were raised
+            assert len(war) >= 1
+
+            # Test the warning messages, ensuring each attribute is present
+            found_war = pysat.instruments.methods.testing.eval_dep_warnings(
+                war, [self.warn_msg])
+
+            for fwar in found_war:
+                assert fwar, "didn't find warning about: {:}".format(self.warn_msg)
+
+        return