+++ CHANGELOG.md
@@ -23,13 +23,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
       - the pysatMadrigal: jro_isr, dmsp_ivm
       - the pysatSpaceWeather: sw_dst, sw_f107, sw_kp
       - the pysatModels: ucar_tiegcm
+      - the pysatNASA: cnofs_ivm, cnofs_plp, cnofs_vefi, de2_lang, de2_nacs,
+        de2_rpa, de2_wats, icon_euv, icon_fuv, icon_ivm, icon_might, iss_fpmu,
+        omni_hro, rocsat1_ivm, sport_ivm, timed_saber, timed_see
    - The usage of a numeric string for `sat_id` to specify number of points
      in test instruments
-   - SpaceWeather, Incubator (DEMETER), and Madrigal instrument method
+   - SpaceWeather, Incubator (DEMETER), Madrigal, CDAWeb, and ICON instrument
+     methods
+   - The DE2 and ICON Constellation objects
 - Documentation
    - Updated docstrings with deprecation notes
 - Bug Fix
    - Closes files after download in NASA CDAWeb methods
+   - Updated the ICON constellation object to reflect new instrument standards
 - Added version cap for numpy
 
 ## [2.2.2] - 2020-11-23